requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-bootstrap==0.2.4
django-bootstrap4==1.0.1
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
gunicorn==20.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==4.1.4
